+++ spec/jwt/claims/not_before_spec.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+RSpec.describe JWT::Claims::NotBefore do
+  let(:payload) { { 'nbf' => (Time.now.to_i + 5) } }
+
+  describe '#validate!' do
+    context 'when nbf is in the future' do
+      it 'raises JWT::ImmatureSignature' do
+        expect { described_class.new(leeway: 0).validate!(context: Struct.new(:payload).new(payload)) }.to raise_error JWT::ImmatureSignature
+      end
+    end
+
+    context 'when nbf is in the past' do
+      let(:payload) { { 'nbf' => (Time.now.to_i - 5) } }
+
+      it 'does not raise error' do
+        expect { described_class.new(leeway: 0).validate!(context: Struct.new(:payload).new(payload)) }.not_to raise_error
+      end
+    end
+
+    context 'when leeway is given' do
+      it 'does not raise error' do
+        expect { described_class.new(leeway: 10).validate!(context: Struct.new(:payload).new(payload)) }.not_to raise_error
+      end
+    end
+  end
+end