+++ lib/jwt/jwa/hmac_rbnacl_fixed.rb
@@ -3,8 +3,12 @@
 module JWT
   module Algos
     module HmacRbNaClFixed
+      include JWT::JWA::SignatureAlgorithm
+
       MAPPING   = { 'HS512256' => ::RbNaCl::HMAC::SHA512256 }.freeze
       SUPPORTED = MAPPING.keys
+
+      register_algorithm(*SUPPORTED)
 
       class << self
         def sign(algorithm, msg, key)