+++ lib/jwt/jwa/unsupported.rb
@@ -3,16 +3,16 @@
 module JWT
   module JWA
     module Unsupported
-      module_function
+      include JWT::JWA::SignatureAlgorithm
 
-      SUPPORTED = [].freeze
+      class << self
+        def sign(*)
+          raise NotImplementedError, 'Unsupported signing method'
+        end
 
-      def sign(*)
-        raise NotImplementedError, 'Unsupported signing method'
-      end
-
-      def verify(*)
-        raise JWT::VerificationError, 'Algorithm not supported'
+        def verify(*)
+          raise JWT::VerificationError, 'Algorithm not supported'
+        end
       end
     end
   end